# documentation/api-docs.yaml
openapi: 3.0.3
info:
  title: Zilla Backend
  description: |-
    Zilla Backend OpenAPI 3.0 specification.
  # termsOfService: http://swagger.io/terms/
  contact:
    email: urosvceh@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /api/user/login:
    post:
      security: []
      summary: Login
      description: Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "schemas/login.schema.yaml"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "schemas/bearer.token.schema.yaml"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Invalid Password
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: Email Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "422":
          description: Token Validation Failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/user/logout:
    post:
      summary: Logout
      description: Logout
      operationId: logout
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized Access
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
  /api/user/set-forgotten-password:
    post:
      security: []
      summary: Set a new password
      description: Set a new password
      operationId: setForgottenPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                securityCode:
                  type: string
                newPassword:
                  type: string
              required:
                - token
                - newPassword
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "schemas/bearer.token.schema.yaml"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Unauthorized Access
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "422":
          description: Token Validation Failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  #################### admin
...